removed safeMath for 8.0 contracts + updated tests

+++ packages/contracts/contracts/mocks/MockCurveThreepool.sol
@@ -67,7 +67,7 @@ contract MockCurveThreepool {
     uint256 slippage = (_token_amount * withdrawalSlippageBps) / 10000;
     uint256 transferOut = _token_amount - slippage;
 
-    uint256 idx = uint256(i);
+    uint128 idx = uint128(i);
     tokens[idx].approve(address(this), transferOut);
     tokens[idx].mint(address(this), transferOut);
     tokens[idx].transferFrom(address(this), msg.sender, transferOut);
